+++ ci/before_script.sh
@@ -5,8 +5,11 @@ set -e
 # load travis-cargo
 pip install 'travis-cargo<0.2' --user
 
+test -x $HOME/.cargo/bin/cargo-install-update || cargo install cargo-update
+test -x $HOME/.cargo/bin/rustfmt || cargo install rustfmt
+cargo install-update rustfmt
+
 if test "$TRAVIS_RUST_VERSION" = "nightly"; then
-    if ! cargo install --list | grep -q clippy; then
-        cargo install clippy
-    fi
+    test -x $HOME/.cargo/bin/cargo-clippy || cargo install clippy
+    cargo install-update clippy
 fi